Move debug printing logic to separate class
